+++ core/src/main/java/org/uma/jmetal/util/grid/imp/VariationAdaptiveGrid.java
@@ -79,7 +79,7 @@ public class VariationAdaptiveGrid<S extends Solution<?>>
      * @return The number of solutions into a specific hypercube.
      */
     @Override
-    public int getLocationDensity(int location) {
+    public double getLocationDensity(int location) {
         return densityMemory[location];
     }
